+++ Source/Core/Core/PowerPC/SignatureDB.cpp
@@ -96,8 +96,7 @@ void SignatureDB::List()
 	{
 		INFO_LOG(OSHLE, "%s : %i bytes, hash = %08x", entry.second.name.c_str(), entry.second.size, entry.first);
 	}
-	INFO_LOG(OSHLE, "%lu functions known in current database.",
-		(unsigned long)database.size());
+	INFO_LOG(OSHLE, "%zu functions known in current database.", database.size());
 }
 
 void SignatureDB::Clear()
